runpod_build/requirements.txt
# AutoCortes RunPod Serverless - Requirements
# Gerado automaticamente

# RunPod
runpod>=1.3.0

# Core ML/AI
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# Whisper (Transcrição)
openai-whisper>=20230314
faster-whisper>=0.7.0

# LLM (Llama/DeepSeek)
llama-cpp-python>=0.2.0
huggingface-hub>=0.16.0

# Video Processing
moviepy>=1.0.3
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Audio Processing
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.1

# Utilities
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Image Processing
scikit-image>=0.21.0

# Streamlit (para testes locais)
streamlit>=1.28.0

# Google Gemini (fallback)
google-generativeai>=0.3.0

# Outros
colorama>=0.4.6
